requirements.txt
pip-upgrader==1.4.15
pyaml==20.3.1
termcolor==1.1.0
